requirements.txt
streamlit==1.29.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
chromadb==0.4.22
pypdf==6.0.0
tiktoken==0.5.2
python-dotenv==1.0.0
requests==2.31.0
openai==1.6.1
volcengine-python-sdk
